# docs/normative/v0.6/contracts/AI-05_OPENAPI.yaml
openapi: 3.1.0
info:
  title: Paquetería Local Culiacán API
  version: 0.6.0
  description: Contrato REST normativo v0.6. REST/PostgreSQL son autoridad; SignalR distribuye eventos confirmados. Seguridad
    tenant, tracking público y outbox siguen contratos ejecutables.
servers:
- url: /api/v1
security:
- oidc: []
...
paths:
...
  /orders:
    post:
      tags:
      - Orders
      operationId: createOrder
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    get:
      tags:
      - Orders
      operationId: listOrders
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/OrderStatus'
      - name: owner_org_id
        in: query
        schema:
          type: string
          format: uuid
      - name: cursor
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Tenant-filtered orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  next_cursor:
                    type:
                    - string
                    - 'null'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /orders/{orderId}:
    get:
      tags:
      - Orders
      operationId: getOrder
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '404':
          $ref: '#/components/responses/UniformNotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /orders/{orderId}/transitions:
    post:
      tags:
      - Orders
      operationId: transitionOrder
      parameters:
      - $ref: '#/components/parameters/OrganizationContext'
      - $ref: '#/components/parameters/OrderId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitionRequest'
      responses:
        '200':
          description: Transition applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/.well-known/openid-configuration
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    OrganizationContext:
      name: X-Organization-Id
      in: header
      required: true
      description: Organización activa; debe corresponder a una membresía ACTIVE del usuario.
      schema:
        type: string
        format: uuid
  responses:
...
    Conflict:
      description: State, idempotency or concurrency conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated actor lacks capability inside an otherwise visible tenant context
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UniformNotFound:
      description: Missing, cross-tenant, expired or revoked resource. Response must not reveal existence.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ServiceUnavailable:
      description: >-
        Required provider or infrastructure dependency is unavailable. Every
        authenticated operation can report it while the identity context
        cannot be resolved, and every tenant operation while the mandatory
        PLATFORM_ADMIN tenant-activation audit cannot be written
        (AI05-DECLARE-EMITTED-ERRORS).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Money:
      type: object
      required:
      - currency
      - amount_cents
      properties:
        currency:
          type: string
          const: MXN
        amount_cents:
          type: integer
          minimum: 0
          format: int64
    OrderStatus:
      type: string
      enum:
      - DRAFT
      - CONFIRMED
      - READY_FOR_PICKUP
      - ASSIGNED
      - AT_PICKUP
      - PICKED_UP
      - IN_TRANSIT
      - DELIVERING
      - FAILED_ATTEMPT
      - RESCHEDULED
      - RETURNING
      - RETURNED
      - DELIVERED
      - CLOSED
      - CLAIM_OPEN
      - CLAIM_RESOLVED
      - CANCELLED
...
    CreateOrderRequest:
      type: object
      required:
      - quote_id
      - payer_type
      - acceptance
      properties:
        quote_id:
          type: string
          format: uuid
        payer_type:
          type: string
          enum:
          - SENDER
          - RECIPIENT
          - BUSINESS_ACCOUNT
        acceptance:
          type: object
          required:
          - terms_version
          - privacy_version
          - accepted_at
          - acceptance_channel
          properties:
            terms_version:
              type: string
              minLength: 1
              maxLength: 64
              description: AI05-INPUT-LIMITS; the bound CSV-001 already applies per row. A longer value is rejected with 409 before any order is created.
            privacy_version:
              type: string
              minLength: 1
              maxLength: 64
              description: AI05-INPUT-LIMITS; the bound CSV-001 already applies per row. A longer value is rejected with 409 before any order is created.
            accepted_at:
              type: string
              format: date-time
              description: Client-observed acceptance timestamp; persisted as accepted_at_client after UTC normalization.
            acceptance_channel:
              type: string
              enum:
              - WEB
              - PWA
              - ASSISTED
              - API
    Order:
      type: object
      required:
      - id
      - public_id
      - owner_org_id
      - status
      - price_net
      - version
      - origin_location_id
      - destination_location_id
      - service_type
      - pricing_tier
      - total
      - city_id
      properties:
        id:
          type: string
          format: uuid
        public_id:
          type: string
        owner_org_id:
          type: string
          format: uuid
        operator_org_id:
          type:
          - string
          - 'null'
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        price_net:
          $ref: '#/components/schemas/Money'
        version:
          type: integer
          minimum: 1
        origin_location_id:
          type: string
          format: uuid
        destination_location_id:
          type: string
          format: uuid
        service_type:
          type: string
          enum:
          - SAME_DAY
          - URGENT
          - SCHEDULED_ROUTE
        quote_id:
          type: string
          format: uuid
        city_id:
          type: string
          format: uuid
        service_area_id:
          type:
          - string
          - 'null'
          format: uuid
        pricing_tier:
          $ref: '#/components/schemas/PricingTier'
        total:
          $ref: '#/components/schemas/Money'
        claim_window_ends_at:
          type:
          - string
          - 'null'
          format: date-time
        finalized_at:
          type:
          - string
          - 'null'
          format: date-time
    OrderDetail:
      allOf:
      - $ref: '#/components/schemas/Order'
      - type: object
        properties:
          timeline:
            type: array
            items:
              $ref: '#/components/schemas/OrderEvent'
    TransitionRequest:
      type: object
      required:
      - target_status
      - reason
      - expected_version
      properties:
        target_status:
          $ref: '#/components/schemas/OrderStatus'
        reason:
          type: string
          maxLength: 500
        expected_version:
          type: integer
          minimum: 1
        metadata:
          type: object
          additionalProperties: true
...
    OrderEvent:
      type: object
      required:
      - event_type
      - occurred_at
      properties:
        event_type:
          type: string
        occurred_at:
          type: string
          format: date-time
    Problem:
      type: object
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
...
    PricingTier:
      type: string
      enum:
      - OCCASIONAL
      - BUSINESS_1_49
      - BUSINESS_50_199
      - BUSINESS_200_499
      - BUSINESS_500_PLUS
      - CUSTOM
...